requirements2.txt
absl-py==0.15.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
atomicwrites==1.4.0
attrs==21.2.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
build==0.10.0
CacheControl==0.12.13
cachetools==4.2.4
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.7
cleo==2.0.1
click==8.0.3
cloudpickle==2.0.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
configparser==5.1.0
crashtest==0.4.1
cycler==0.11.0
debugpy @ file:///C:/ci/debugpy_1637073815078/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
distlib==0.3.4
docker-pycreds==0.4.0
dulwich==0.21.5
dunamai==1.7.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
ffmpeg==1.4
filelock==3.4.2
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.41.1
gym==0.23.1
-e git+https://github.com/Farama-Foundation/MicroRTS-Py.git@5b24b0203ac38bf33259846f3fa78969b59e0a05#egg=gym_microrts
gym-notices==0.0.8
html5lib==1.1
identify==2.4.5
idna==3.3
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1687138353019/work
iniconfig==1.1.1
installer==0.7.0
ipykernel @ file:///D:/bld/ipykernel_1655369313836/work
ipython @ file:///D:/bld/ipython_1680185618122/work
jaraco.classes==3.2.3
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
Jinja2==3.0.2
JPype1==1.3.0
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1687700988094/work
jupyter_core @ file:///D:/bld/jupyter_core_1686775775850/work
kiwisolver==1.3.2
lockfile==0.12.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
more-itertools==9.1.0
msgpack==1.0.5
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
nodeenv==1.6.0
numpy==1.21.4
oauthlib==3.1.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==1.3.4
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathtools==0.1.2
peewee==3.14.8
PettingZoo==1.16.0
pexpect==4.8.0
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==8.4.0
pkginfo==1.9.6
pkgutil_resolve_name==1.3.10
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1687705014305/work
pluggy==1.0.0
poetry==1.5.1
poetry-core==1.6.1
poetry-dynamic-versioning==0.13.1
poetry-plugin-export==1.4.0
pre-commit==2.17.0
promise==2.3
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
protobuf==3.19.1
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
ptyprocess==0.7.0
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1681904169130/work
pyparsing==2.4.7
pyproject_hooks==1.0.0
pytest==6.2.5
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2021.3
pywin32==305.1
PyYAML==5.4.1
pyzmq @ file:///C:/b/abs_655zk4a3s8/croot/pyzmq_1686601465034/work
rapidfuzz==2.15.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==1.0.0
rsa==4.7.2
sentry-sdk==1.4.3
shellingham==1.5.0.post1
shortuuid==1.0.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
stable-baselines3==1.1.0
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
subprocess32==3.5.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
toml==0.10.2
tomlkit==0.7.2
torch==1.7.1
tornado @ file:///D:/bld/tornado_1656937958156/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
trove-classifiers==2023.5.24
trueskill==0.4.5
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1687998981955/work
urllib3==1.26.7
virtualenv==20.13.0
wandb==0.12.6
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
webencodings==0.5.1
Werkzeug==2.0.2
yaspin==2.1.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
